Add qwerty diagram
This describes my layout for the 3-row keyboard qwerty mode.  There are
lots of combos, which works well, as the keys are designed to be pressed
this way.  There are also three layers for numbers, navigation, and
function keys.

The number layer also adds keys for hex, with 'e' in a weird place,
because the regular 'e' key is taken by a '3'.